+++ packages/types/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.2.4 - _TBD, 2018_
+
+    * Add `data` to `TxData` (#413)
+    * Add `number` as an option to `ContractEventArg` (#413)
+
 ## v0.2.1 - _February 9, 2018_
 
     * Fix publishing issue where .npmignore was not properly excluding undesired content (#389)