+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.4.0 - _TBD, 2018_
+
+    * Use `ethers-contracts` as a backend to decode event args (#413)
+
 ## v0.3.2 - _February 9, 2018_
 
     * Fix publishing issue where .npmignore was not properly excluding undesired content (#389)